models/VulBERTa-CNN_VulBERTa-MLP/code/requirements.txt
absl-py==1.0.0
alembic==1.4.1
appdirs==1.4.4
argon2-cffi==20.1.0
ase==3.20.1
astunparse==1.6.3
async-generator==1.10
attrs==20.2.0
azure-core==1.8.2
azure-storage-blob==12.5.0
backcall==0.2.0
beautifulsoup4==4.10.0
bertviz==1.2.0
bleach
boto3==1.20.5
botocore==1.23.5
bs4==0.0.1
cachetools==4.2.4
captum==0.4.1
certifi==2021.10.8
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cloudpickle==1.6.0
configparser==5.0.1
cpgqls-client==0.0.7
cryptography==3.2.1
cssselect==1.1.0
cycler==0.10.0
dacite==1.6.0
databricks-cli==0.13.0
dataclasses==0.6
datasets
decorator==4.4.2
deepspeed
defusedxml==0.6.0
Deprecated==1.2.12
dill==0.3.4
docker==4.3.1
docker-pycreds==0.4.0
entrypoints==0.3
fake-useragent==0.1.11
filelock==3.0.12
Flask==1.1.2
flatbuffers==2.0
flawfinder==2.0.18
fsspec==2021.7.0
future==0.18.2
gast==0.4.0
gensim==3.8.3
gitdb==4.0.5
GitPython==3.1.11
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googledrivedownloader==0.4
gorilla==0.3.0
grpcio==1.42.0
gunicorn==20.0.4
h5py==3.1.0
huggingface-hub
idna==2.10
importlib-metadata==4.8.2
intervaltree==3.1.0
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
isodate==0.6.0
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.17.0
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.6.3
jupyterlab==2.2.9
jupyterlab-pygments==0.1.2
jupyterlab-server==1.2.0
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
libclang==12.0.0
llvmlite==0.35.0
lxml==4.6.2
Mako==1.1.3
Markdown==3.3.6
MarkupSafe==1.1.1
matplotlib==3.3.3
mistune==0.8.4
mlflow==1.11.0
msrest==0.6.19
multiprocess==0.70.12.2
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.2
networkx==2.5
ninja==1.10.2.2
notebook==6.1.4
numba==0.52.0
numexpr==2.7.3
numpy==1.19.2
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.4
pandas==1.3.2
pandocfilters==1.4.3
parallelformers==1.0.1
parse==1.19.0
parso==0.7.1
pathtools==0.1.2
pexpect==4.8.0
phply==1.2.5
pickleshare==0.7.5
Pillow==8.0.1
ply==3.11
prometheus-client==0.8.0
prometheus-flask-exporter==0.18.1
promise==2.3
prompt-toolkit==3.0.8
protobuf==3.13.0
psutil==5.7.3
ptyprocess==0.6.0
pyarrow==5.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyee==8.2.2
PyGithub==1.55
Pygments==2.7.2
PyJWT==2.1.0
PyNaCl==1.4.0
pyparsing==2.4.7
pyppeteer==1.0.2
pyquery==1.4.3
pyrsistent==0.17.3
python-dateutil==2.8.1
python-editor==1.0.4
python-louvain==0.14
pytorch-nlp==0.5.0
pytz==2020.1
PyYAML==5.3.1
pyzmq==19.0.2
querystring-parser==1.2.4
rdflib==5.0.0
regex==2020.11.13
requests==2.24.0
requests-html==0.10.0
requests-oauthlib==1.3.0
rsa==4.7.2
s3transfer==0.5.0
sacremoses==0.0.43
scikit-learn==0.23.2
scipy==1.5.3
Send2Trash==1.5.0
sentencepiece==0.1.96
sentry-sdk==0.19.4
shortuuid==1.0.1
six==1.15.0
sklearn==0.0
smart-open==3.0.0
smmap==3.0.4
sortedcontainers==2.3.0
soupsieve==2.3.1
SQLAlchemy==1.3.13
sqlparse==0.4.1
subprocess32==3.5.4
tables==3.6.1
tabulate==0.8.7
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorboardX==1.8
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.22.0
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
threadpoolctl==2.1.0
tokenizers==0.10.1
torch==1.7.0
torch-cluster==1.5.9
torch-geometric==1.6.3
torch-scatter==2.0.5
torch-sparse==0.6.8
torch-spline-conv==1.2.0
torchaudio==0.7.0
torchtext==0.8.0
torchvision==0.8.1
tornado==6.0.4
tqdm==4.51.0
traitlets==5.0.5
transformers==4.4.1
transformers-interpret==0.5.2
triton==1.1.1
typing==3.7.4.3
typing-extensions==3.7.4.3
urllib3==1.25.11
w3lib==1.22.0
wandb==0.10.12
watchdog==0.10.4
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.57.0
websockets==10.2
Werkzeug==1.0.1
wrapt==1.12.1
xxhash==2.0.2
zipp==3.6.0
